fixup move MapResourcesToKubernetesResources to stage package
